pyproject.toml
[project]
name = "autoware-onnx-test-ui"
version = "0.1.0"
description = "Python UI for testing Autoware ONNX ML models without a full ROS2/Autoware setup"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.35",
    "onnxruntime>=1.18",
    "opencv-python>=4.9",
    "numpy>=1.26",
    "pillow>=10.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.4",
]
